Now returns all useful metrics from LuxorAPI to prometheus

+++ src/main/java/pro/zackpollard/luxorexporter/api/types/Hashrate.java
@@ -8,6 +8,5 @@ import lombok.Data;
 public class Hashrate {
 
     private double hashrate;
-    private long time;
-
+    private double time;
 }
